Build provisioner cleanup

+++ provisioners/cleanup.sh
@@ -14,7 +14,7 @@ rm /var/lib/dhcp/*;
 
 # Clean up udev rules.
 # Make sure Udev doesn't block our network.
-rm /etc/udev/rules.d/70-persistent-net.rules;
+[[ -r /etc/udev/rules.d/70-persistent-net.rules ]] && rm /etc/udev/rules.d/70-persistent-net.rules;
 mkdir /etc/udev/rules.d/70-persistent-net.rules;
 rm -rf /dev/.udev/;
 rm /lib/udev/rules.d/75-persistent-net-generator.rules;
